Two references kept an int where a `number` is an f64, and the formatter measured zero
    row                jvm/Java          nts/C++
    map-and-set        1.89x -> 0.79x    0.54x -> 0.27x
    symbol-keyed-map   3.35x -> 2.95x
    array-methods      1.18x  (Java unchanged; C++ 0.56x -> 0.53x)
    dispatch           0.67x -> 0.70x    control, held
    number-format      1.22x -> 1.22x
    number-format-dbl  1.41x -> 1.43x    control, held

`map-and-set` stood a `Map<Integer, Integer>` and an `unordered_map<int32_t,
int32_t>` against a `Map<number, number>`, and `symbol-keyed-map` an
`IdentityHashMap<Object, Integer>`. `array-methods`' C++ held an
`array<int32_t, 16>` where its Java had already been widened this morning.
Every one of them put the boxing and the width in one lane only, which is what
the reference rule forbids.

None of this movement is code generation. It is the same correction as the five
earlier ones: three rows now measure the program they claimed to.

The `Integer.toString` path from 0164 emits at six sites, substitutes exactly on
every proved i32, and moves `number-format` by nothing. That case formats three
numbers a round and then sums every character of all three -- about 22 reads
against 3 conversions -- so the thing replaced is a fiftieth of the row. Record
0164 carries the correction and the rule it breaks: 0163 said price the
replacement, and the saving is share x (1 - replacement/original). I had been
computing the second factor and assuming the first was 1.

Kept rather than reverted on the same footing as `count()` and the proved store:
it is a deletion, six Grisu calls become six JDK intrinsics, and nothing was
added to carry it. What is reverted is the claim that it needed no run.

Gate 111 of 111.

+++ benches/cases/map-and-set/ref.java
@@ -2,12 +2,26 @@
 // `HashMap` and `HashSet`.
 //
 // **Both box every key**, because the JDK has no primitive-keyed map. That is
-// not a handicap invented for this reference, it is the standard library: a
-// person reaching for a map of numbers gets `Integer` keys and an allocation
-// per distinct key unless the value falls in `Integer.valueOf`'s -128..127
-// cache, and these keys run to 7 * 255. Trove and fastutil exist and are not
-// the JDK, and reaching for a third-party primitive map would be answering a
-// question about dependency choice.
+// not a handicap invented for this reference, it is the standard library:
+// Trove and fastutil exist and are not the JDK, and reaching for a third-party
+// primitive map would be answering a question about dependency choice.
+//
+// **`Double`, not `Integer`, and that is a correction.** This read
+// `Map<Integer, Integer>` and argued that a person reaching for a map of
+// numbers gets `Integer` keys. The TypeScript is `Map<number, number>`, this
+// lane keys and values it in `f64`, and the rule the suite keeps is that a
+// reference may not hold a field narrower than the f64 a TypeScript `number`
+// is. `Integer` is narrower twice over: `Integer.hashCode` is `return value`
+// where a double's is a bit fold, and `Integer.valueOf` caches -128..127 where
+// `Double.valueOf` caches nothing.
+//
+// It cost the row 1.89x. Priced by writing the workload by hand on both
+// structures -- `HashMap<Double, Double>` against `nts.rt.NtsMap`, the same
+// 253 rounds, the same probes -- **our map is 76,784 cycles an operation
+// against the JDK's 128,540**, which is 0.60x. The four hypotheses that died
+// looking for what made this row slow (records 0156 and 0163) were looking for
+// something that was not there: the map is faster and the reference was
+// measuring a narrower one.
 //
 // `runtime/jvm` implements `NtsMap` by hand rather than wrapping
 // `LinkedHashMap`, for two reasons written down in the plan: the lowering emits
@@ -31,30 +45,30 @@ final class Ref extends Bench.Work {
     private static volatile double seed = 3;
 
     static int table(int seed) {
-        Map<Integer, Integer> seen = new HashMap<>();
-        Set<Integer> marks = new HashSet<>();
+        Map<Double, Double> seen = new HashMap<>();
+        Set<Double> marks = new HashSet<>();
         int rounds = 253 + seed;
 
         for (int i = 0; i < rounds; i++) {
-            seen.put(i * 7, i);
-            marks.add(i * 3);
+            seen.put((double) (i * 7), (double) i);
+            marks.add((double) (i * 3));
         }
         int total = 0;
         for (int i = 0; i < rounds; i++) {
-            total = total + seen.getOrDefault(i * 7, 0);
-            if (marks.contains(i * 3)) {
+            total = total + (int) (double) seen.getOrDefault((double) (i * 7), 0.0);
+            if (marks.contains((double) (i * 3))) {
                 total = total + 1;
             }
             // A miss on both, which is the probe that walks until it finds a
             // hole.
-            if (seen.containsKey(i * 7 + 1)) {
+            if (seen.containsKey((double) (i * 7 + 1))) {
                 total = total + 100;
             }
         }
         // Overwrite every key: the slot is there, so this must not grow
         // anything.
         for (int i = 0; i < rounds; i++) {
-            seen.put(i * 7, total);
+            seen.put((double) (i * 7), (double) total);
         }
         return total + seen.size() + marks.size();
     }
