+++ benches/cases/symbol-keyed-map/ref.java
@@ -20,19 +20,26 @@ final class Ref extends Bench.Work {
 
     static int work(int seed) {
         final int step = seed;
-        final IdentityHashMap<Object, Integer> events = new IdentityHashMap<>();
-        events.put(A, 1);
-        events.put(B, 2);
-        events.put(C, 3);
-        events.put(D, 4);
+        // `Double` values, not `Integer`. The TypeScript is
+        // `Map<symbol, number>` and this lane holds the value in an `f64`, so
+        // an `Integer` is a field narrower than the `number` it stands for --
+        // and narrower in the way that matters here, since `Integer.valueOf`
+        // caches -128..127 and every value in this table is inside that cache
+        // while `Double.valueOf` caches nothing. The *keys* are symbols on both
+        // sides and are matched.
+        final IdentityHashMap<Object, Double> events = new IdentityHashMap<>();
+        events.put(A, 1.0);
+        events.put(B, 2.0);
+        events.put(C, 3.0);
+        events.put(D, 4.0);
         int total = 0;
         for (int i = 0; i < 4096; i++) {
             final int which = (i ^ step) & 3;
             final Object key = which == 0 ? A : which == 1 ? B : which == 2 ? C : D;
-            final Integer found = events.get(key);
-            total = total + (found == null ? 0 : found);
-            final Integer missing = events.get(ABSENT);
-            total = total + (missing == null ? 0 : missing);
+            final Double found = events.get(key);
+            total = total + (int) (found == null ? 0.0 : found);
+            final Double missing = events.get(ABSENT);
+            total = total + (int) (missing == null ? 0.0 : missing);
         }
         return total;
     }
